move TestMergeTrafficPolicy
